Implemented listview.builder for cards
I was using a janky for loop to generate the button widgets from the dummy data but all the tutorials i was seeing were starting off from a "listview.builder" widget that handles everything for you. I went ahead and redid some decently large chunks of code to do it the "proper' way. Not connected to the firebase yet but should be a couple lines of code from here now.