Added invite functionality
added the ability to invite instead of users being "forced" into being a caretaker. The user now recieves an invite they can either accept or decline. The invite is stored in a seperate table called "invites". The addbaby and invites are now found in a sub menu navigated to by the plus sign on the main menu