cleaned up caretaker additions
made the user input validation more responsive with a popup instead of a command line message. Also does things like clear the text line if the input was good and give a response saying "invite sent".

The column in baby_menu was also changed to a listview to get rid of the pixel overflow