The edit bio and edit profile pic are added
The emergency ID now has a bio and profile pic that can be edited. The bio is stored in the firebase baby doc and the pic is stored as a link to a file in the firebase storage. This should allow anyone with access to the baby to see the same pic and bio.